% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyHelper.R
\name{applySGolayFilter}
\alias{applySGolayFilter}
\title{apply a sgolay filter (WIP)}
\usage{
applySGolayFilter(df, continuousVar, k_05 = 10)
}
\arguments{
\item{df}{- a potentially grouped data frame with continuous observations}

\item{continuousVar}{- the column containing the samples}

\item{k_05}{- the half value of the filter width}
}
\description{
apply a sgolay filter (WIP)
}
