% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.R
\name{bloodResultsSimulation}
\alias{bloodResultsSimulation}
\title{tidy dataframe of the simulation of blood test results with known distributions for individual outcomes.}
\usage{
bloodResultsSimulation(n, seed = 101)
}
\description{
tidy dataframe of the simulation of blood test results with known distributions for individual outcomes.
}
