% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyHelper.R
\name{byRow}
\alias{byRow}
\title{A row wise dataframe iterator in for loops}
\usage{
byRow(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
a list of lists
}
\description{
Converts a dataframe into a list of lists
}
