% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyContinuousEntropy.R
\name{calculateContinuousEntropy}
\alias{calculateContinuousEntropy}
\title{calculate entropy of a sequence of continuous variables}
\usage{
calculateContinuousEntropy(df, continuousVar, method, ...)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different types of continuous variable}

\item{method}{- the method employed - valid options are "MontgomerySmith", "Histogram", "Grassberger", "InfoTheo", "Compression"}

\item{...}{- the other parameters are passed onto the implementations}

\item{groupVars}{- the columns of the discrete value quoted by the vars() function (e.g. ggplot facet_wrap)}
}
\value{
a single value for the entropy of the vector
}
\description{
calculate entropy of a sequence of continuous variables
}
