% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyContinuousEntropy.R
\name{calculateContinuousEntropy_PDF}
\alias{calculateContinuousEntropy_PDF}
\title{calculate differential entropy of a continuous value (X) using a probability function approach:}
\usage{
calculateContinuousEntropy_PDF(
  df,
  continuousVar,
  probabilityMethod = "Kernel",
  ...
)
}
\arguments{
\item{df}{- may be grouped, in which case the grouping is interpreted as different types of discrete variable}

\item{continuousVar}{- the column of the continuous value (Y)}

\item{probabilityMethod}{- the method of calculating a PDF - valid options are "SGolay", "Kernel"}

\item{...}{- passed onto probabilitiesFromContinuous}
}
\value{
a dataframe containing the disctinct values of the groups of df, and for each group an entropy value (H). If df was not grouped this will be a single entry
}
\description{
calculate differential entropy of a continuous value (X) using a probability function approach:
}
