% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyEntropy.R
\name{calculateContinuousEntropy_Quantile}
\alias{calculateContinuousEntropy_Quantile}
\title{calculate differential entropy of a continuous value (X) using a quantile function smoothing approach:
Entropy of continuous data - e.g. SGolay approach / discretisation}
\usage{
calculateContinuousEntropy_Quantile(
  df,
  continuousVar,
  k_05 = 10,
  collect = FALSE,
  ...
)
}
\arguments{
\item{df}{- may be grouped, in which case the grouping is interpreted as different types of discrete variable}

\item{continuousVar}{- the column of the continuous value (Y)}

\item{k_05}{- the half window width of the SG filter that smooths the data. This is dependent on data but typically not less that 10.}

\item{collect}{- if TRUE will collect dbplyr tables before processing, otherwise (the default) will fail on dbplyr tables}

\item{groupVars}{- the columns of the discrete value quoted by the vars() function (e.g. ggplot facet_wrap)}
}
\value{
a dataframe containing the disctinct values of the groups of df, and for each group an entropy value (H). If df was not grouped this will be a single entry
}
\description{
The Savitsky Golay filter width required to make this at all accurate needs a reasonable amount of data ~ 20 points
S. M. Sunoj and P. G. Sankaran, “Quantile based entropy function,” Stat. Probab. Lett., vol. 82, no. 6, pp. 1049–1053, Jun. 2012, doi: 10.1016/j.spl.2012.02.005. [Online]. Available: http://www.sciencedirect.com/science/article/pii/S0167715212000521
}
