% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyTFIDF.R
\name{calculateCooccurrence}
\alias{calculateCooccurrence}
\title{Calculate a co-occurrence matrix for concepts associated with a grouping}
\usage{
calculateCooccurrence(
  groupedDf,
  conceptIdVar,
  countVar = NULL,
  filteredIdfDf = NA,
  bigResult = FALSE
)
}
\arguments{
\item{groupedDf}{a dataframe whose grouping defines the "document"}

\item{conceptIdVar}{a field that contains the unique id of a "term"}

\item{countVar}{a field that contains a count. If this is given then it is assumed that the concept & document combinations are unique}

\item{filteredIdfDf}{an optional data frame containing idf information from this or another corpus. It is recommended that this is a heavily filtered list as the result will be the square of the size of this}

\item{bigResult}{are you expecting a big result?}
}
\value{
a data frame with co-occurrence stats for each concept in filteredIdf (i.e. document)

the co-occurrence matrix as a dataframe
}
\description{
The grouped dataframe here acts as a "Document" from the perpective of the co-occurrence but might be a person
}
