% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscreteBinaryMI.R
\name{calculateDiscreteBinaryMI}
\alias{calculateDiscreteBinaryMI}
\title{calculate mutual information between a discrete value (X) with a binary outcome.}
\usage{
calculateDiscreteBinaryMI(df, discreteVars, countVar = NA, ...)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different types of continuous variable - the grouping may be w.g. a test or concept.}

\item{discreteVars}{- the column(s) of the categorical value (X) quoted by vars(...)}

\item{countVar}{- optional the column of the count variable - how often does the event happen? If missing then this will be the assumed to be individual observations. In this case the df is a contingency table}

\item{...}{- the other parameters are passed onto the implementations}
}
\value{
a dataframe containing the disctinct values of the groups of df, and for each group a mutual information column (I). If df was not grouped this will be a single entry
}
\description{
calculate mutual information between a discrete value (X) with a binary outcome.
}
