% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscreteContinuousMI.R
\name{calculateDiscreteContinuousConditionalMI_Kernel}
\alias{calculateDiscreteContinuousConditionalMI_Kernel}
\title{calculate pointwise mutual information between a categorical value (X) and a continuous value (Y). I.e. the self information of X conditioned on each of the possible values for X}
\usage{
calculateDiscreteContinuousConditionalMI_Kernel(
  df,
  discreteVars,
  continuousVar,
  collect = FALSE,
  ...
)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different types of continuous variable}

\item{discreteVars}{- the column(s) of the categorical value (X) quoted by vars(...)}

\item{continuousVar}{- the column of the continuous value (Y)}

\item{collect}{- if TRUE will collect dbplyr tables before processing, otherwise (the default) will fail on dbplyr tables}
}
\value{
a dataframe containing the disctinct values of the groups of df, and for each group a mutual information column (I). If df was not grouped this will be a single entry
}
\description{
calculate pointwise mutual information between a categorical value (X) and a continuous value (Y). I.e. the self information of X conditioned on each of the possible values for X
}
