% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscreteContinuousMI.R
\name{calculateDiscreteContinuousMI_Compression}
\alias{calculateDiscreteContinuousMI_Compression}
\title{calculate mutual information between a categorical value (X) and a continuous value (Y) using a compression algorithm based on.}
\usage{
calculateDiscreteContinuousMI_Compression(
  df,
  discreteVar,
  continuousVar,
  collect = FALSE,
  ...
)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different types of continuous variable}

\item{discreteVar}{- the column of the categorical value (X)}

\item{continuousVar}{- the column of the continuous value (Y)}

\item{collect}{- unless TRUE this function will fail on dbplyr tables as there is no SQL implementation}
}
\value{
a dataframe containing the disctinct values of the groups of df, and for each group a mutual information column (I). If df was not grouped this will be a single entry
}
\description{
Universal and accessible entropy estimation using a compression algorithm Ram Avinery, Micha Kornreich, Roy Beck https://arxiv.org/abs/1709.10164
}
