% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscreteContinuousMI.R
\name{calculateDiscreteContinuousMI_DiscretiseByValue}
\alias{calculateDiscreteContinuousMI_DiscretiseByValue}
\title{calculate mutual information between a categorical value (X) and a continuous value (Y) using a discretisation and infotheo}
\usage{
calculateDiscreteContinuousMI_DiscretiseByValue(
  df,
  discreteVar,
  continuousVar,
  bins = NA,
  binStrategy = linearBySize(slope = 8, minBins = 4, maxBins = 100),
  discreteMethod = "Histogram",
  ...
)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different types of continuous variable}

\item{discreteVar}{- the column of the categorical value (X)}

\item{continuousVar}{- the column of the continuous value (Y)}

\item{bins}{- the number of bins}

\item{binStrategy}{- the a function to genereate the bins on a group by group basis}

\item{discreteMethod}{- What methd will be used to calculate the MI once discretised?}
}
\value{
a dataframe containing the disctinct values of the groups of df, and for each group a mutual information column (I). If df was not grouped this will be a single entry
}
\description{
calculate mutual information between a categorical value (X) and a continuous value (Y) using a discretisation and infotheo
}
