% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscreteContinuousMI.R
\name{calculateDiscreteContinuousMI_PDF}
\alias{calculateDiscreteContinuousMI_PDF}
\title{calculate mutual information between a categorical value (X) and a continuous value (Y) using an estimator of PDF}
\usage{
calculateDiscreteContinuousMI_PDF(
  df,
  discreteVars,
  continuousVar,
  probabilityMethod = "Kernel",
  ...
)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different types of continuous variable}

\item{discreteVars}{- the column(s) of the categorical value (X) quoted by vars(...)}

\item{continuousVar}{- the column of the continuous value (Y)}

\item{collect}{- if TRUE will collect dbplyr tables before processing, otherwise (the default) will fail on dbplyr tables}
}
\value{
a dataframe containing the disctinct values of the groups of df, and for each group a mutual information column (I). If df was not grouped this will be a single entry
}
\description{
calculate mutual information between a categorical value (X) and a continuous value (Y) using an estimator of PDF
}
