% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscreteContinuousMI.R
\name{calculateDiscreteContinuousMI_SGolay}
\alias{calculateDiscreteContinuousMI_SGolay}
\title{calculate mutual information between a categorical value (X) and a continuous value (Y)}
\usage{
calculateDiscreteContinuousMI_SGolay(
  df,
  discreteVar,
  continuousVar,
  k_05 = 10,
  collect = FALSE,
  ...
)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different types of continuous variable}

\item{discreteVar}{- the column of the categorical value (X)}

\item{continuousVar}{- the column of the continuous value (Y)}

\item{k_05}{- the half window width of the SG filter that smooths the data. This is dependent on data but typically not less that 10.}

\item{collect}{- if TRUE will collect dbplyr tables before processing, otherwise (the default) will fail on dbplyr tables}
}
\value{
a dataframe containing the disctinct values of the groups of df, and for each group a mutual information column (I). If df was not grouped this will be a single entry
}
\description{
calculate mutual information between a categorical value (X) and a continuous value (Y)
}
