% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactoredCode.R
\name{calculateDiscreteContinuousPointwiseConditionalEntropy_Entropy}
\alias{calculateDiscreteContinuousPointwiseConditionalEntropy_Entropy}
\title{calculate conditional mutual information between a discrete value (X) and a continuous value (Y) using estimates of differential entropy}
\usage{
calculateDiscreteContinuousPointwiseConditionalEntropy_Entropy(
  df,
  discreteVars,
  continuousVar,
  entropyMethod = "Quantile",
  ...
)
}
\arguments{
\item{df}{- may be grouped, in which case the grouping is interpreted as different types of discrete variable}

\item{discreteVars}{- the column(s) of the categorical value (X) quoted by vars(...)}

\item{continuousVar}{- the column of the continuous value (Y)}

\item{entropyMethod}{- the method used to calculate the entropy (see ?tidyinfostats::calculateDiscreteEntropy) - defaults to "Grassberger"}

\item{collect}{- unless TRUE this function will fail on dbplyr tables as there is no SQL implementation}
}
\value{
a dataframe containing the disctinct values of the groups of df, and for each group a mutual information column (I). If df was not grouped this will be a single entry
}
\description{
calculate conditional mutual information between a discrete value (X) and a continuous value (Y) using estimates of differential entropy
}
