% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscreteDiscreteMI.R
\name{calculateDiscreteDiscreteMI_Empirical}
\alias{calculateDiscreteDiscreteMI_Empirical}
\title{calculate mutual information between a discrete value (X) and a discrete value (Y)}
\usage{
calculateDiscreteDiscreteMI_Empirical(
  df,
  groupXVar,
  groupYVar,
  adjust = TRUE,
  ...
)
}
\arguments{
\item{df}{- may be grouped, in which case the grouping is interpreted as different types of discrete variable}

\item{groupXVar}{- the column of the discrete value (X)}

\item{groupYVar}{- the column of the discrete value (Y)}

\item{adjust}{- Apply a miller-madow adjustement to the result (default: TRUE)}
}
\value{
a dataframe containing the distinct values of the groups of df, and for each group a mutual information column (I). If df was not grouped this will be a single entry
}
\description{
calculate mutual information between a discrete value (X) and a discrete value (Y)
}
