% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscreteDiscreteMI.R
\name{calculateDiscreteDiscreteMI_Entropy}
\alias{calculateDiscreteDiscreteMI_Entropy}
\title{calculate mutual information between a discrete value (X) and a discrete value (Y) using estimates of entropy}
\usage{
calculateDiscreteDiscreteMI_Entropy(
  df,
  groupXVar,
  groupYVar,
  entropyMethod = "Grassberger",
  ...
)
}
\arguments{
\item{df}{- may be grouped, in which case the grouping is interpreted as different types of discrete variable}

\item{groupXVar}{- the column of the discrete value (X)}

\item{groupYVar}{- the column of the discrete value (Y)}

\item{entropyMethod}{- the method used to calculate the entropy (see ?tidyinfostats::calculateEntropy) - defaults to "Grassberger"}
}
\value{
a dataframe containing the disctinct values of the groups of df, and for each group a mutual information column (I). If df was not grouped this will be a single entry
}
\description{
calculate mutual information between a discrete value (X) and a discrete value (Y) using estimates of entropy
}
