% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscreteBinaryMI.R
\name{calculateDiscretePresentValuesMI}
\alias{calculateDiscretePresentValuesMI}
\title{calculate mutual information between a categorical value (X) and its presence in a data set.}
\usage{
calculateDiscretePresentValuesMI(
  df,
  discreteVars,
  sampleVars,
  sampleCount = NULL,
  sampleCountDf = NULL,
  ...
)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different types of continuous variable}

\item{discreteVars}{- the column(s) of the categorical value (X) quoted by vars(...) (e.g. outcome)}

\item{sampleVars}{- the column(s) of the sample identifier. From this we expect (at least) one outcome per sample}

\item{sampleCount}{- an integer containing the count of all samples per outcome (discreteVar)}

\item{sampleCountDf}{- a dataframe containing columns for grouping, and discreteVars, N and N_x columns with counts
see expectSamplesByOutcome(...)}
}
\value{
a dataframe containing the distinct values of the groups of df, and for each group a mutual information column (I). If df was not grouped this will be a single entry
}
\description{
This calculates the mutual information of a feature not being present in all samples
}
\details{
This is relevant for sparse data sets with many features such as NLP terms, where a term as a
feature is only flagged as present in the samples.
}
