% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscreteSelfInformation.R
\name{calculateDiscreteSelfInformation}
\alias{calculateDiscreteSelfInformation}
\title{calculate self information of a discrete variable}
\usage{
calculateDiscreteSelfInformation(df, discreteVars, method = "Histogram", ...)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different types of discrete variable}

\item{discreteVars}{- the columns of the discrete value quoted by the vars() function (e.g. ggplot facet_wrap)}

\item{method}{- the method employed - valid options are "Histogram"}

\item{...}{- the other parameters are passed onto the implementations}
}
\value{
a dataframe containing the distinct values of the discrete variable, per group, with columns I_x, and method
}
\description{
calculate self information of a discrete variable
}
