% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyEntropy.R
\name{calculateEntropy}
\alias{calculateEntropy}
\title{calculate entropy of a sequence of}
\usage{
calculateEntropy(df, groupVars, method, ...)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different types of continuous variable}

\item{groupVars}{- the columns of the discrete value quoted by the vars() function (e.g. ggplot facet_wrap)}

\item{method}{- the method employed - valid options are "MontgomerySmith", "Histogram", "Grassberger", "InfoTheo", "Compression"}

\item{...}{- the other parameters are passed onto the implementations}
}
\value{
a single value for the entropy of the vector
}
\description{
calculate entropy of a sequence of
}
