% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyEntropy.R
\name{calculateEntropy_Compression}
\alias{calculateEntropy_Compression}
\title{calculate mutual information between a categorical value (X) and a continuous value (Y) using a compression algorithm based on.}
\usage{
calculateEntropy_Compression(
  df,
  groupVars,
  orderingVar = NULL,
  collect = FALSE,
  ...
)
}
\arguments{
\item{df}{- may be grouped, in which case the grouping is interpreted as different types of discrete variable}

\item{groupVars}{- the column of the discrete value (X)}

\item{orderingVar}{- (optional) the column of an ordering variable (e.g. time) - if missing assumes df order,}

\item{collect}{- if TRUE will collect dbplyr tables before processing, otherwise (the default) will fail on dbplyr tables}
}
\value{
a dataframe containing the disctinct values of the groups of df, and for each group an entropy value (H). If df was not grouped this will be a single entry
}
\description{
Universal and accessible entropy estimation using a compression algorithm Ram Avinery, Micha Kornreich, Roy Beck https://arxiv.org/abs/1709.10164
}
