% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyEntropy.R
\name{calculateEntropy_Histogram}
\alias{calculateEntropy_Histogram}
\title{calculate entropy of an optionally discrete value (X) using a histogram approach}
\usage{
calculateEntropy_Histogram(df, groupVars, mm = TRUE, ...)
}
\arguments{
\item{df}{- may be grouped, in which case the grouping is interpreted as different types of discrete variable}

\item{groupVars}{- the columns of the discrete value quoted by the vars() function (e.g. ggplot facet_wrap)}

\item{mm}{- Apply a miller-madow adjustment to the result}
}
\value{
a dataframe containing the disctinct values of the groups of df, and for each group an entropy value (H). If df was not grouped this will be a single entry
}
\description{
calculate entropy of an optionally discrete value (X) using a histogram approach
}
