% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyEntropy.R
\name{calculateEntropy_MontgomerySmith}
\alias{calculateEntropy_MontgomerySmith}
\title{calculate entropy of an optionally ordered discrete value (X) using estimates of entropy from method 2 in the following:}
\usage{
calculateEntropy_MontgomerySmith(df, groupVars, orderingVar = NULL, ...)
}
\arguments{
\item{df}{- may be grouped, in which case the grouping is interpreted as different types of discrete variable}

\item{groupVars}{- the columns of the discrete value quoted by the vars() function (e.g. ggplot facet_wrap)}

\item{orderingVar}{- (optional) the column of an ordering variable (e.g. time) - if missing assumes df order,}
}
\value{
a dataframe containing the disctinct values of the groups of df, and for each group an entropy value (H). If df was not grouped this will be a single entry
}
\description{
S. Montgomery-Smith and T. Schürmann, “Unbiased Estimators for Entropy and Class Number,” arXiv, 18-Oct-2014. Available: http://arxiv.org/abs/1410.5002
}
\details{
This is not particularly useful as requires large sample size before it becomes accurate.
}
