% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscreteDiscreteMI.R
\name{calculateMultiClassMI}
\alias{calculateMultiClassMI}
\title{Calculate single mutual information score from multiclass groups in dplyr friendly manner.}
\usage{
calculateMultiClassMI(df, adjust = TRUE)
}
\arguments{
\item{df}{a dataframe containing one observation per row & minimally p_x1y1, p_x1, p_y1 columns (see probabilitiesFromCounts)}

\item{adjust}{apply a Miller-Madow adjustment to the result}
}
\value{
the datatable with additional columns for MI
}
\description{
The purpose of this is to make it possible to calculate MI from tidy data. This is useful where you have a a data from that
represents a multi-class confusion matrix with unique combinations of inputs and probabilities for the co-occurrence and marginal probabilities 
already calculated. Typically this will be generated by the probabilitiesFromGroups function.
}
