% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscreteSelfInformation.R
\name{calculateSelfInformation_Histogram}
\alias{calculateSelfInformation_Histogram}
\title{calculate the self information for all values of a discrete variable in a dplyr friendly way}
\usage{
calculateSelfInformation_Histogram(
  df,
  groupVars,
  countVar = NULL,
  mm = TRUE,
  ...
)
}
\arguments{
\item{df}{- may be grouped, in which case the grouping is interpreted as different types of discrete variable}

\item{groupVars}{- the columns of the discrete value quoted by the vars() function (e.g. ggplot facet_wrap)}

\item{countVar}{- (optional) if this datafram represents summary counts, the columns of the summary variable.}

\item{mm}{- Apply a miller-madow adjustment to the result? default = TRUE}
}
\value{
a dataframe containing the distinct values of the discrete variable, per group, with columns I_x, and method
}
\description{
see: R. de Matos Simoes and F. Emmert-Streib, “Influence of statistical estimators of mutual information and data heterogeneity on the inference of gene regulatory networks,” PLoS One, vol. 6, no. 12, p. e29279, Dec. 2011 [Online]. Available: http://dx.doi.org/10.1371/journal.pone.0029279
}
