% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyTFIDF.R
\name{calculateTfidf}
\alias{calculateTfidf}
\title{Calculate a term frequecy / inverse document frequency for concepts associated with a grouping}
\usage{
calculateTfidf(
  groupedDf,
  conceptIdVar,
  countVar = NULL,
  idfDf = NULL,
  k1 = 1.2,
  b = 0.95
)
}
\arguments{
\item{groupedDf}{a dataframe whose grouping defines the "document"}

\item{conceptIdVar}{a field that contains the unique id of a "term"}

\item{countVar}{a field that contains a count. If this is given then it is assumed that the concept & document combinations are unique}

\item{idfDf}{an optional data frame containing idf information from this or another corpus}

\item{k1}{default 1.2 - okapi BM25 parameter}

\item{b}{default 0.95 - okapi BM25}
}
\value{
a data frame with tfidf stats for each concept in each group (i.e. document)
}
\description{
The grouped dataframe here acts as a "Document" from the perpective of the TFIDF calculation but might be a person
TODO: fix methods of calculating tfidf
}
