% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyHelper.R
\name{coefficientAdj}
\alias{coefficientAdj}
\title{adjust the coefficients of a SGoly filter (WIP)}
\usage{
coefficientAdj(sgolayTable, sampleSizeVar)
}
\arguments{
\item{sgolayTable}{- a data frame containing the sgolay coefficients joined to a data set}

\item{sampleSizeVar}{- the colum containing the sample size it is to be applied to}
}
\description{
adjust the coefficients of a SGoly filter (WIP)
}
