% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyHelper.R
\name{collectAsSparseMatrix}
\alias{collectAsSparseMatrix}
\title{Converts a tidy dataframe into a Matrix::sparseMatrix}
\usage{
collectAsSparseMatrix(
  df,
  rowVar,
  colVar,
  valueVar = NULL,
  rowNameVar = NULL,
  colNameVar = NULL,
  ...
)
}
\arguments{
\item{df}{- a df}

\item{rowVar}{- the dataframe columns(s) which define the matrix row, quoted by vars(...) - typically this is the observation id}

\item{colVar}{- the dataframe columns(s) which define the matrix columns, quoted by vars(...) - typically this is the feature id}

\item{valueVar}{- the name of the value variable. (#TODO could be missing - in which case use binary)}

\item{rowNameVar}{- (optional) the dataframe column continaing the row names otherwise use rowVar}

\item{colNameVar}{- (optional) the dataframe column continaing the column names otherwise use colVar}

\item{...}{- other parameters passes to Matrix::sparseMatrix}
}
\value{
a dbplyr dataframe containing the grouped function
}
\description{
offloading the majority of processing onto sql if dbplyr tables are involved
}
