% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyHelper.R
\name{collectAsTrainingSet}
\alias{collectAsTrainingSet}
\title{Converts a tidy dataframe into a Matrix::sparseMatrix of features & associated outcome vector}
\usage{
collectAsTrainingSet(
  df,
  sampleVar,
  outcomeVar,
  featureVar,
  valueVar = NULL,
  featureNameVar = NULL,
  factorise = TRUE,
  ...
)
}
\arguments{
\item{df}{- a df}

\item{sampleVar}{- the dataframe columns(s) which define the matrix row, quoted by vars(...) - typically this is the observation id ( see createSequentialIdentifier(...) )}

\item{outcomeVar}{- the dataframe columns(s) which define the matrix columns, quoted by vars(...) - typically this is the feature id}

\item{featureVar}{- the dataframe columns(s) which define the matrix columns, quoted by vars(...) - typically this is the feature id}

\item{valueVar}{- the name of the value variable. (#TODO could be missing - in which case use binary)}

\item{...}{- other parameters passes to Matrix::sparseMatrix & as.factor (for outcomes)}

\item{rowNameVar}{- (optional) the dataframe column continaing the row names otherwise use rowVar}

\item{colNameVar}{- (optional) the dataframe column continaing the column names otherwise use colVar}
}
\value{
a dbplyr dataframe containing the grouped function
}
\description{
offloading the majority of processing onto sql if dbplyr tables are involved.
}
