% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyHelper.R
\name{collectOutcomeVector}
\alias{collectOutcomeVector}
\title{Converts a tidy dataframe into a Matrix::sparseMatrix}
\usage{
collectOutcomeVector(df, rowVar, outcomeVar, factorise = TRUE, ...)
}
\arguments{
\item{df}{- a df}

\item{rowVar}{- the dataframe columns(s) which define the matrix row, quoted by vars(...) - typically this is the observation id}

\item{outcomeVar}{- (optional) the dataframe column continaing the outcome}

\item{factorise}{- convert outcomeVar to a factor (default FALSE)}

\item{...}{- other parameters passed to as.factor}
}
\value{
a dbplyr dataframe containing the grouped function
}
\description{
offloading the majority of processing onto sql if dbplyr tables are involved
}
