% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyHelper.R
\name{createSequentialIdentifier}
\alias{createSequentialIdentifier}
\title{Create an id based for each sample based on a unique combination of variables}
\usage{
createSequentialIdentifier(df, groupVars, sampleIdVar)
}
\arguments{
\item{df}{- a df which may be grouped in which case the grouping can be interpreted as a feature and each group must have the same number in it.}

\item{groupVars}{- the columns(s) for which we want to create a label - as a list of columns quoted by vars(...). Essentially a combination of discrete variables. i.e. identifies each sample.}

\item{sampleIdVar}{- the name of the target column for the new discrete variable (or sample identifier)}
}
\value{
a dbplyr dataframe containing the grouped function
}
\description{
Sometime simpler to rename multiple grouping columns as one single column which encodes all the possible combinations.
}
