% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscretise.R
\name{cutsDfFromVector}
\alias{cutsDfFromVector}
\title{generate a data fram of cuts for every group in a dataframe from a vector of cuts, and the dataframe}
\usage{
cutsDfFromVector(df, cuts)
}
\arguments{
\item{df}{- a grouped dataframe}

\item{cuts}{- a vector of cuts - not including the -Inf, Inf ends}
}
\value{
a datafram of cuts
}
\description{
generate a data fram of cuts for every group in a dataframe from a vector of cuts, and the dataframe
}
