% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyHelper.R
\docType{data}
\name{digammaLookup}
\alias{digammaLookup}
\title{a set of the first 50 digamma values}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50 rows and 2 columns.}
\usage{
digammaLookup
}
\description{
a set of the first 50 digamma values
}
\keyword{datasets}
