% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscretise.R
\name{discretise}
\alias{discretise}
\title{calculate entropy of a sequence of}
\usage{
discretise(df, continuousVar, discreteOutputVar, method, ...)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different types of continuous variable}

\item{continuousVar}{- the columns that define the discrete subgroups of the data.}

\item{discreteOutputVar}{- the name of the value to create in the dataframe}

\item{method}{- the method employed - valid options are "ByRank", "ByValue", "Manual"}

\item{...}{- the other parameters are passed onto the implementations}
}
\value{
a single value for the entropy of the vector
}
\description{
calculate entropy of a sequence of
}
