% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscretise.R
\name{discretise_ByRank}
\alias{discretise_ByRank}
\title{Discretise data by the rank of a continuous variable}
\usage{
discretise_ByRank(
  df,
  continuousVar,
  discreteOutputVar,
  bins = NA,
  binStrategy = linearBySize(slope = 8, min = 4, max = 100),
  ...
)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different types of continuous variable}

\item{continuousVar}{- the columns that define the continuous data.}

\item{discreteOutputVar}{- the name of the value to create in the dataframe for the discrete data}

\item{bins}{- (optional) number of bins}

\item{binStrategy}{- if the number of bins is not set they will be calculated using this bin strategy}

\item{...}{- other options passed onto tidyinfostats::discretise_Manual}
}
\value{
a dataframe with an additional column with the discrete categories
}
\description{
Discretise data by the rank of a continuous variable
}
