% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscretise.R
\name{discretise_ByValue}
\alias{discretise_ByValue}
\title{Discretise data by the value of a continuous variable}
\usage{
discretise_ByValue(
  df,
  continuousVar,
  discreteOutputVar,
  bins = NA,
  binStrategy = linearBySize(slope = 8, minBins = 4, maxBins = 100),
  ...
)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different types of continuous variable}

\item{continuousVar}{- the columns that define the continuous data.}

\item{discreteOutputVar}{- the name of the value to create in the dataframe for the discrete data}

\item{bins}{- (optional) number of bins}

\item{binStrategy}{- if the number of bins is not set they will be calculated using this bin strategy}
}
\value{
a dataframe with an additional column with the discrete categories
}
\description{
Discretise data by the value of a continuous variable
}
