% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscretise.R
\name{discretise_Manual}
\alias{discretise_Manual}
\title{Discretise data using pre-defined break points}
\usage{
discretise_Manual(
  df,
  continuousVar,
  discreteOutputVar,
  cutsDf,
  lowerBounded = FALSE,
  factorise = FALSE,
  noUnicode = ("tbl_sql" \%in\% class(df)),
  ...
)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different types of continuous variable}

\item{continuousVar}{- the columns that define the discrete subgroups of the data.}

\item{discreteOutputVar}{- the name of the value to create in the dataframe}

\item{cutsDf}{- manual break points as a dataframe containing cut points for every group (see cutsDfFromVector).}

\item{lowerBounded}{- default FALSE, should the lower or upper values be included in the group}

\item{factorise}{- convert discrete values into an ordered factor (alternative is a character string). This is only useful if you have a single set of cuts for all groups.}

\item{noUnicode}{- by default unicode characters are not used for the label if the target is a dbplyr table}
}
\value{
a dataframe containing the discreteOutputVar column
}
\description{
Discretise data using pre-defined break points
}
