% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscreteContinuousMI.R
\name{expectedObservations}
\alias{expectedObservations}
\title{summary info}
\usage{
expectedObservations(df, discreteVars, expected)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different types of continuous variable}

\item{discreteVars}{- the column(s) of the categorical value (X) quoted by vars(...)}

\item{expected}{- the total number of samples for each outcome expected if there were no missing values.}
}
\description{
summary info
}
