% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValues.R
\name{expectFixedSamples}
\alias{expectFixedSamples}
\title{specify the number of expected observations for each class based on  a fixed number.}
\usage{
expectFixedSamples(df, discreteVars, sampleCount, ...)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different features}

\item{discreteVars}{- the column(s) of the categorical value (X) quoted by vars(...)}

\item{sampleCount}{- the number of samples per outcome (equal numbers).}
}
\description{
specify the number of expected observations for each class based on  a fixed number.
}
