% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValues.R
\name{expectOnePerSample}
\alias{expectOnePerSample}
\title{specify the number of expected observations for each class based on a grouping of the data}
\usage{
expectOnePerSample(df, discreteVars, sampleVars, ...)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different features}

\item{discreteVars}{- the column(s) of the categorical value (X) quoted by vars(...) i.e. typically the outcome}

\item{sampleVars}{- the column(s) that identify the sample space, quoted by vars(...) i.e. usually the sample identifier}
}
\description{
specify the number of expected observations for each class based on a grouping of the data
}
