% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValues.R
\name{expectSamplesByOutcome}
\alias{expectSamplesByOutcome}
\title{specify the number of expected observations for each class based on  a fixed number.}
\usage{
expectSamplesByOutcome(df, samplesPerOutcomeDf, ...)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different types of continuous variable}

\item{discreteVars}{- the column(s) of the categorical value (X) quoted by vars(...) (usually the outcome)}

\item{samplePerOutcomeDf}{- a dataframe containing one row per outcome and the number of samples expected for that outcome in column N}
}
\description{
specify the number of expected observations for each class based on  a fixed number.
}
