% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorisedCalculations.R
\name{fBeta}
\alias{fBeta}
\title{calculate mutual information from probabilities - see calculateConfusionMatrixStats to generate these probabilities from count information}
\usage{
fBeta(p_x1y1, p_x0y1, p_x1y0, beta = 1)
}
\arguments{
\item{p_x1y1}{probability of cooccurrence of x and y - true positives.}

\item{p_x0y1}{probability of occurrence of y without x - false positives}

\item{p_x1y0}{probability of occurrence of x without y - false negatives}

\item{betathe}{beta coefficient for the f score}
}
\value{
the fBeta score
}
\description{
in the terminology here x is prediction and y is outcome / gold standard
}
