% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyKNN.R
\name{findKNN}
\alias{findKNN}
\title{calculates an approximate KNN for sourceDf entries in targetDf entries}
\usage{
findKNN(
  sourceDf,
  targetDf,
  sourceIdVar,
  targetIdVar,
  k = 4,
  matchVars = vars(long, lat),
  distanceVar = "distance",
  rankVar = "rank",
  ...
)
}
\arguments{
\item{sourceDf}{a dataframe with a unique Id column}

\item{targetDf}{a dataframe with a unique Id column}

\item{sourceIdVar}{the unique Id column of the sourceDf}

\item{targetIdVar}{the unique Id column of the targetDf}

\item{k}{the k of knn}

\item{matchVars}{the columns to calculate the distance on - must be numerics and present in both sourceDf and targetDf, escaped by vars(...)}

\item{distanceVar}{what will the distance measure be named?}

\item{...}{other parameters passed onto RANN::nn2()}
}
\value{
a dataframe containing sourceIdVar, targetIdVar, distanceVar and k columns
}
\description{
relise on RANN library for actual calculation. Simply reformats the results into a mapping table format
}
