% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscretise.R
\name{fixedWidth}
\alias{fixedWidth}
\title{A binning strategy generator. Creates a function that can be used to create discretisation cuts based on the statistical paramaters of the data to be discretised
in particular the minimum and maximum value of the observed data (in a group by group fashion)}
\usage{
fixedWidth(width)
}
\arguments{
\item{width}{- the width of each bin}
}
\description{
A binning strategy generator. Creates a function that can be used to create discretisation cuts based on the statistical paramaters of the data to be discretised
in particular the minimum and maximum value of the observed data (in a group by group fashion)
}
