% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyHelper.R
\name{group_mutate}
\alias{group_mutate}
\title{Some databases don't support window functions over grouped data. This requires a workaround to group and summarise then join the data}
\usage{
group_mutate(df, ...)
}
\arguments{
\item{df}{- a df which may be grouped, in which case the value is interpreted as different types of continuous variable}

\item{...}{- the contents of the mutuate function}
}
\value{
a dbplyr dataframe containing the grouped function
}
\description{
Some databases don't support window functions over grouped data. This requires a workaround to group and summarise then join the data
}
