% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionFunctions.R
\name{KumaraswamyDistribution}
\alias{KumaraswamyDistribution}
\title{Get a kumaraswamy distribution wrapper}
\description{
Get a kumaraswamy distribution wrapper

Get a kumaraswamy distribution wrapper
}
\details{
This is a Kumaraswamy distrbution with support from x=0..1
}
\keyword{distributions}
\section{Super class}{
\code{\link[tidyinfostats:Distribution]{tidyinfostats::Distribution}} -> \code{KumaraswamyDistribution}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mode}}{the mode of the kumaraswamy}

\item{\code{iqr}}{the iqr of the normal distribuition}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{KumaraswamyDistribution$new()}}
\item \href{#method-label}{\code{KumaraswamyDistribution$label()}}
\item \href{#method-clone}{\code{KumaraswamyDistribution$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="getInverseCdf">}\href{../../tidyinfostats/html/Distribution.html#method-getInverseCdf}{\code{tidyinfostats::Distribution$getInverseCdf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="getPdf">}\href{../../tidyinfostats/html/Distribution.html#method-getPdf}{\code{tidyinfostats::Distribution$getPdf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="inverse">}\href{../../tidyinfostats/html/Distribution.html#method-inverse}{\code{tidyinfostats::Distribution$inverse()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="p">}\href{../../tidyinfostats/html/Distribution.html#method-p}{\code{tidyinfostats::Distribution$p()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="plot">}\href{../../tidyinfostats/html/Distribution.html#method-plot}{\code{tidyinfostats::Distribution$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="q">}\href{../../tidyinfostats/html/Distribution.html#method-q}{\code{tidyinfostats::Distribution$q()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="sample">}\href{../../tidyinfostats/html/Distribution.html#method-sample}{\code{tidyinfostats::Distribution$sample()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="theoreticalEntropy">}\href{../../tidyinfostats/html/Distribution.html#method-theoreticalEntropy}{\code{tidyinfostats::Distribution$theoreticalEntropy()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="twoDp">}\href{../../tidyinfostats/html/Distribution.html#method-twoDp}{\code{tidyinfostats::Distribution$twoDp()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Kumaraswamy distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KumaraswamyDistribution$new(mode = runif(1, 0.1, 0.9), iqr = runif(0.1, 0.3))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mode}}{the mode of the distribution}

\item{\code{iqr}}{the iqr of the target (must be strictly less than 0.5)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-label"></a>}}
\subsection{Method \code{label()}}{
get a label for this distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KumaraswamyDistribution$label()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KumaraswamyDistribution$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
