% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscretise.R
\name{linearBySize}
\alias{linearBySize}
\title{A binning strategy generator. Creates a function that can be used to create discretisation cuts based on the statistical paramaters of the data to be discretised
in particular the sample size, minimum and maximum value of the observed data (in a group by group fashion)}
\usage{
linearBySize(slope, minBins, maxBins, fn = fixedNumber)
}
\arguments{
\item{slope}{- the number of observations for each bin if the observation numbers fall between minBins*slope and maxBins*slope}

\item{minBins}{- the smallest number of bins to create for each group}

\item{maxBins}{- the largest number of bins to create for each group}

\item{fn}{- the distribution function that genertes the cut points (e.g. tidyinfostats::fixedNumber (uniform), tidyinfostats::logNormalCentiles  )}
}
\description{
A binning strategy generator. Creates a function that can be used to create discretisation cuts based on the statistical paramaters of the data to be discretised
in particular the sample size, minimum and maximum value of the observed data (in a group by group fashion)
}
