% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscretise.R
\name{logNormalCentiles}
\alias{logNormalCentiles}
\title{A binning strategy generator. Creates a function that can be used to create discretisation cuts based on the statistical paramaters of the data to be discretised
in particular the mean and sd of the observed data (in a group by group fashion)}
\usage{
logNormalCentiles(bins)
}
\arguments{
\item{bins}{- the number of bins to create out of the centiles of a log normal distribution}
}
\description{
A binning strategy generator. Creates a function that can be used to create discretisation cuts based on the statistical paramaters of the data to be discretised
in particular the mean and sd of the observed data (in a group by group fashion)
}
