% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionFunctions.R
\name{LogNormalDistribution}
\alias{LogNormalDistribution}
\title{Get a log normal distribution wrapper}
\description{
Get a log normal distribution wrapper

Get a log normal distribution wrapper
}
\keyword{distributions}
\section{Super class}{
\code{\link[tidyinfostats:Distribution]{tidyinfostats::Distribution}} -> \code{LogNormalDistribution}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mu}}{the mean of the normal distribuition}

\item{\code{sigma}}{the sd of the normal distribuition}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LogNormalDistribution$new()}}
\item \href{#method-label}{\code{LogNormalDistribution$label()}}
\item \href{#method-clone}{\code{LogNormalDistribution$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="getInverseCdf">}\href{../../tidyinfostats/html/Distribution.html#method-getInverseCdf}{\code{tidyinfostats::Distribution$getInverseCdf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="getPdf">}\href{../../tidyinfostats/html/Distribution.html#method-getPdf}{\code{tidyinfostats::Distribution$getPdf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="inverse">}\href{../../tidyinfostats/html/Distribution.html#method-inverse}{\code{tidyinfostats::Distribution$inverse()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="p">}\href{../../tidyinfostats/html/Distribution.html#method-p}{\code{tidyinfostats::Distribution$p()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="plot">}\href{../../tidyinfostats/html/Distribution.html#method-plot}{\code{tidyinfostats::Distribution$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="sample">}\href{../../tidyinfostats/html/Distribution.html#method-sample}{\code{tidyinfostats::Distribution$sample()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="theoreticalEntropy">}\href{../../tidyinfostats/html/Distribution.html#method-theoreticalEntropy}{\code{tidyinfostats::Distribution$theoreticalEntropy()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
plot this dictributions as pdf and cdf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogNormalDistribution$new(mode = runif(1, 1, 4), sd = runif(1, 0.5, 5))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mode}}{the mode - must be 1 or greater TODO: find out why}

\item{\code{sd}}{the sd}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-label"></a>}}
\subsection{Method \code{label()}}{
gets a label for this distribution based on the parameters passed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogNormalDistribution$label()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogNormalDistribution$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
