% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorisedCalculations.R
\name{mi}
\alias{mi}
\title{calculate mutual information from probabilities}
\usage{
mi(p_x1y1, p_x0y1, p_x1y0)
}
\arguments{
\item{p_x1y1}{probability of cooccurrence of x and y.}

\item{p_x0y1}{probability of occurrence of y without x}

\item{p_x1y0}{probability of occurrence of x without y}
}
\value{
the mutual information
}
\description{
calculate mutual information from probabilities
}
