% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorisedCalculations.R
\name{npmi}
\alias{npmi}
\title{A calculate normalised pointwise mutual information from probabilities}
\usage{
npmi(p_xy, p_x, p_y)
}
\arguments{
\item{p_xy}{probability of cooccurrence of x and y.}

\item{p_x}{probability of occurrence of x}

\item{p_y}{probability of occurrence of y}
}
\value{
the normalised pointwise mutual information
}
\description{
A calculate normalised pointwise mutual information from probabilities
}
