% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorisedCalculations.R
\name{pmi}
\alias{pmi}
\title{A calculate pointwise mutual information from probabilities}
\usage{
pmi(p_xy, p_x, p_y)
}
\arguments{
\item{p_xy}{probability of cooccurrence of x and y.}

\item{p_x}{probability of occurrence of x}

\item{p_y}{probability of occurrence of y}
}
\value{
the pointwise mutual information
}
\description{
A calculate pointwise mutual information from probabilities
}
