% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorisedCalculations.R
\name{pmiFromCounts}
\alias{pmiFromCounts}
\title{A calculate pointwise mutual information from counts}
\usage{
pmiFromCounts(x1y1, x1, y1, total)
}
\arguments{
\item{x1y1}{cooccurrence of x1 and y1 in corpus.}

\item{x1}{occurrence of x1 in corpus}

\item{y1}{occurrence of y1 in corpus}

\item{total}{total size of corpus}
}
\value{
the pointwise mutual information
}
\description{
A calculate pointwise mutual information from counts
}
