% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyConvertToProbabilities.R
\name{probabilitiesFromConfusionMatrix}
\alias{probabilitiesFromConfusionMatrix}
\title{Helper function to calculate probability from confusion matrix stats in dplyr friendly manner}
\usage{
probabilitiesFromConfusionMatrix(df, tpVar, fpVar, fnVar, tnVar)
}
\arguments{
\item{df}{a dataframe containing one observation per row}

\item{tpVar}{the datatable column containing frequency of cooccurrence of true positives}

\item{fpVar}{the datatable column containing frequency of occurrence of false positives}

\item{fnVar}{the datatable column containing frequency of occurrence of false negatives}

\item{tnVar}{the datatable column containing frequency of occurrence of true negatives}
}
\value{
the datatable with additional columns for all the probabilities associated with each outcome (i.e. a 2x2 confusion matrix)
}
\description{
The purpose of this is to calculate the probabilities of binary outcomes from
a table of true pos, true neg, false pos and false neg trials
}
