% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyContinuousProbability.R
\name{probabilitiesFromContinuous}
\alias{probabilitiesFromContinuous}
\title{estimate PDFs from samples in a tidy friendly manner}
\usage{
probabilitiesFromContinuous(df, continuousVar, method = "Kernel", ...)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different types of continuous variable}

\item{continuousVar}{- the column of the continuous value (Y)}

\item{method}{- the method employed - valid options are "Kernel", "SGolay"}

\item{...}{- the other parameters are passed onto the implementations}
}
\value{
a dataframe containing the disctinct values of the groups of df, and for each group a mutual information column (I). If df was not grouped this will be a single entry
}
\description{
estimate PDFs from samples in a tidy friendly manner
}
