% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyContinuousProbability.R
\name{probabilitiesFromContinuous_Kernel}
\alias{probabilitiesFromContinuous_Kernel}
\title{Helper function to calculate probability from continuous data in a tidy friendly manner}
\usage{
probabilitiesFromContinuous_Kernel(
  df,
  continuousVar,
  minVar = NULL,
  maxVar = NULL,
  collect = FALSE
)
}
\arguments{
\item{df}{a dataframe containing a column of a continuous variable X and one row per observation, 
df may also be grouped and in which case the grouping is preserved in the result.}

\item{minVar}{- the name of the column containing the minimum value for the observation / support range}

\item{maxVar}{- the name of the column containing the maximum value for the observation / support range}

\item{continousVar}{- the datatable column(s) containing the observation.}
}
\value{
A mutated datatable with observations of X, the total number of observations of X (N), the probability density (p_x), and self information (I_x) associated with the value of X
}
\description{
The purpose of this is to calculate the probabilities of events from continuous data. 
This function is useful when you have a set of observations from a continuous distribution.
Kernel methods resample the data and produce an evenly spaced output function so needs some sort of support range defined for each group (minVar and maxVar)
}
