% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyConvertToProbabilities.R
\name{probabilitiesFromCounts}
\alias{probabilitiesFromCounts}
\title{Helper function to calculate probability from counts in dplyr friendly manner}
\usage{
probabilitiesFromCounts(df, x1y1Var, x1Var, y1Var, totalVar)
}
\arguments{
\item{df}{a dataframe containing one observation per row}

\item{x1y1Var}{the datatable column containing frequency of cooccurrences of events x1 and y1}

\item{x1Var}{the datatable column containing frequency of occurrences of event x1}

\item{y1Var}{the datatable column containing frequency of occurrences of event y1}

\item{totalVar}{the datatable column containing total number of events}
}
\value{
the datatable with additional columns for all the probabilities associated with each outcome (i.e. a 2x2 confusion matrix)
}
\description{
The purpose of this is to calculate the probabilities of given all binary outcomes from
a table of marginal frequencies in a dplyr friendly way.
}
