% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyHelper.R
\name{sgolayTable}
\alias{sgolayTable}
\title{Precalculates a sgolay filter (WIP)}
\usage{
sgolayTable(polynomialOrder, filterLength, derivative)
}
\arguments{
\item{polynomialOrder}{- order of the poynomial to fit}

\item{filterLength}{- length of the filter}

\item{derivative}{- derivative order for the fitted function}
}
\description{
Precalculates a sgolay filter (WIP)
}
