% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyHelper.R
\name{sparseMatrixToSparseMCsr}
\alias{sparseMatrixToSparseMCsr}
\title{Sparse matrix conversion for liblinear}
\usage{
sparseMatrixToSparseMCsr(X)
}
\arguments{
\item{X}{as Matrix::sparseMatrix}
}
\value{
a SparseM::matrix.csr
}
\description{
Converts a Matrix::sparseMatrix to a SparseM::matrix.csr without blowing up the computer
}
