% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.R
\name{tidyDiscreteUSArrestsCooccurrence}
\alias{tidyDiscreteUSArrestsCooccurrence}
\title{tidy dataframe of the USArrests data with co-occurence of features}
\usage{
tidyDiscreteUSArrestsCooccurrence()
}
\description{
tidy dataframe of the USArrests data with co-occurence of features
}
