% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.R
\name{tidyUSArrestsCooccurrence}
\alias{tidyUSArrestsCooccurrence}
\title{tidy dataframe of the USArrests data with co-occurence of features}
\usage{
tidyUSArrestsCooccurrence()
}
\description{
tidy dataframe of the USArrests data with co-occurence of features
}
