% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyHelper.R
\name{unarrange}
\alias{unarrange}
\title{Unarrange an ordered data set}
\usage{
unarrange(remote_df)
}
\arguments{
\item{remote_df}{- a df which may be a dbplyr table}
}
\value{
a dbplyr dataframe without any ordering
}
\description{
Unarrange an ordered data set
}
