% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionFunctions.R
\name{UniformDistribution}
\alias{UniformDistribution}
\title{Get a uniform distribution wrapper}
\description{
Get a uniform distribution wrapper

Get a uniform distribution wrapper
}
\keyword{distributions}
\section{Super class}{
\code{\link[tidyinfostats:Distribution]{tidyinfostats::Distribution}} -> \code{UniformDistribution}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UniformDistribution$new()}}
\item \href{#method-label}{\code{UniformDistribution$label()}}
\item \href{#method-clone}{\code{UniformDistribution$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="getInverseCdf">}\href{../../tidyinfostats/html/Distribution.html#method-getInverseCdf}{\code{tidyinfostats::Distribution$getInverseCdf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="getPdf">}\href{../../tidyinfostats/html/Distribution.html#method-getPdf}{\code{tidyinfostats::Distribution$getPdf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="inverse">}\href{../../tidyinfostats/html/Distribution.html#method-inverse}{\code{tidyinfostats::Distribution$inverse()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="p">}\href{../../tidyinfostats/html/Distribution.html#method-p}{\code{tidyinfostats::Distribution$p()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="plot">}\href{../../tidyinfostats/html/Distribution.html#method-plot}{\code{tidyinfostats::Distribution$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="sample">}\href{../../tidyinfostats/html/Distribution.html#method-sample}{\code{tidyinfostats::Distribution$sample()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tidyinfostats" data-topic="Distribution" data-id="theoreticalEntropy">}\href{../../tidyinfostats/html/Distribution.html#method-theoreticalEntropy}{\code{tidyinfostats::Distribution$theoreticalEntropy()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
plot this dictributions as pdf and cdf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UniformDistribution$new(min = runif(1, -3, 3), max = min + runif(1, 0.5, 6))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{min}}{the min value of the uniform distribution}

\item{\code{max}}{the max value of the uniform distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-label"></a>}}
\subsection{Method \code{label()}}{
get a label for this distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UniformDistribution$label()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UniformDistribution$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
