% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyHelper.R
\name{calculateDigamma}
\alias{calculateDigamma}
\title{Calculate an estimate of the digamma function value for dplyr and dbplyr dataframes. This will be done in SQL if possible.}
\usage{
calculateDigamma(df, inputVar, outputVar)
}
\arguments{
\item{df}{- a dataframe}

\item{inputVar}{- the reference to the column for which the digamma value will be calculated (this must be an integer)}

\item{outputVar}{- the column that the digamma result will be written to}
}
\description{
Calculate an estimate of the digamma function value for dplyr and dbplyr dataframes. This will be done in SQL if possible.
}
