% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscreteContinuousMI.R
\name{calculateDiscreteContinuousMI}
\alias{calculateDiscreteContinuousMI}
\title{calculate mutual information between a categorical value (X) and a continuous value (Y)}
\usage{
calculateDiscreteContinuousMI(
  df,
  groupXVar,
  valueYVar,
  method = "KWindow",
  ...
)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different types of continuous variable}

\item{groupXVar}{- the column of the categorical value (X)}

\item{valueYVar}{- the column of the continuous value (Y)}

\item{method}{- the method employed - valid options are "KWindow","KNN","SGolay","DiscreteByRank","DiscreteByValue","Compression","Entropy".}

\item{...}{- the other parameters are passed onto the implementations}
}
\value{
a dataframe containing the disctinct values of the groups of df, and for each group a mutual information column (I). If df was not grouped this will be a single entry
}
\description{
calculate mutual information between a categorical value (X) and a continuous value (Y)
}
