% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDiscreteDiscreteMI.R
\name{calculateDiscreteDiscreteMI}
\alias{calculateDiscreteDiscreteMI}
\title{calculate mutual information between a discrete value (X) and a discrete value (Y)}
\usage{
calculateDiscreteDiscreteMI(
  df,
  groupXVar,
  groupYVar,
  method = "Empirical",
  ...
)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different types of continuous variable}

\item{groupXVar}{- the column of the discrete value (X)}

\item{groupYVar}{- the column of the discrete value (Y)}

\item{method}{- the method employed - valid options are "Empirical","MontgomerySmith","Compression","Histogram","Entropy","Grassberger"}

\item{...}{- the other parameters are passed onto the implementations}
}
\value{
a dataframe containing the disctinct values of the groups of df, and for each group a mutual information column (I). If df was not grouped this will be a single entry
}
\description{
calculate mutual information between a discrete value (X) and a discrete value (Y)
}
