% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyTFIDF.R
\name{calculateIdf}
\alias{calculateIdf}
\title{Creates a table with baseline IDF stats for concepts given a set of documents}
\usage{
calculateIdf(groupedDf, conceptIdVar, countVar = NULL, totalDocuments = NA)
}
\arguments{
\item{groupedDf}{a dataframe whose grouping defines the "document"}

\item{conceptIdVar}{a field that contains the unique id of a "term"}

\item{countVar}{a field that contains a count. If this is given then it is assumed that the concept & document combinations are unique}
}
\value{
a data frame with idf stats for each concept in each group (i.e. document)
}
\description{
given a dataframe whose grouping defines the concept to calculate
IDF stats for
}
