% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyHelper.R
\name{collectAsTrainingSet}
\alias{collectAsTrainingSet}
\title{Converts a tidy dataframe into a Matrix::sparseMatrix of features & associated outcome vector}
\usage{
collectAsTrainingSet(
  df,
  sampleVar,
  outcomeVar,
  featureVar,
  valueVar = NULL,
  featureNameVar = NULL,
  factorise = TRUE,
  ...
)
}
\arguments{
\item{df}{- a df}

\item{sampleVar}{- the dataframe columns(s) which define the matrix row, quoted by vars(...) - typically this is the observation id ( see createSequentialIdentifier(...) )}

\item{outcomeVar}{- the dataframe columns(s) which define the matrix columns, quoted by vars(...) - typically this is the feature id}

\item{featureVar}{- the dataframe columns(s) which define the matrix columns, quoted by vars(...) - typically this is the feature id}

\item{valueVar}{- the name of the value variable. (#TODO could be missing - in which case use binary)}

\item{...}{- other parameters passes to Matrix::sparseMatrix & as.factor (for outcomes)}
}
\value{
a list with the following elements:

* rowLabels - the labels for each row in order
* colLabels - the feature labels in order
* matrix - a Matrix::sparseMatrix of the data, with values as doubles
* outcome - a vector of outcomes (probably as a factor)
}
\description{
offloading the majority of processing onto sql if dbplyr tables are involved.
}
