% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyHelper.R
\name{groupMutate}
\alias{groupMutate}
\title{Execute a mutate function on grouped data in all databases}
\usage{
groupMutate(df, ...)
}
\arguments{
\item{df}{- a df which may be a dbplyr table}

\item{...}{- the contents of the mutuate function}
}
\value{
a dbplyr dataframe containing the grouped function
}
\description{
Some databases don't support window functions over grouped data. This requires a workaround to group and summarise then join the data.
All database seem to support group_by(...) and then mutate(x = sum(...)) but not all will do group_by(...) and then mutate(x = mean(...)) for example.
}
\details{
Ranking functions all seem to work (including row_number)

* sum
* row_number
* rank
* lead, lag

Aggregation functions are variably supported by the default mutate depending on the database

* mean
* sd
* min
* max
}
