% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyHelper.R
\name{groupwiseCount}
\alias{groupwiseCount}
\title{A summary function to return a grouped df containing counts (N, N_x)}
\usage{
groupwiseCount(df, groupVars, countVar = NULL, summarise = FALSE)
}
\arguments{
\item{df}{- a df which may be grouped}

\item{groupVars}{- the grouping for which we want to create a label as a list of columns quoted by vars(...)}

\item{outputVar}{- the name of the target column for relabelling}
}
\value{
the grouped dataframe containing the df grouping columns, the groupVars columns, and a groupwise count of both levels of grouping labelled N and N_x and the groupwise p_x.
}
\description{
A summary function to return a grouped df containing counts (N, N_x)
}
