% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyHelper.R
\name{labelGroup}
\alias{labelGroup}
\title{Sometime simpler to rename multiple grouping variables as one single variable which encodes all the possible combinations}
\usage{
labelGroup(df, groupVars, outputVar, summarise = FALSE)
}
\arguments{
\item{df}{- a df which may be grouped}

\item{groupVars}{- the grouping for which we want to create a label as a list of columns quoted by vars(...)}

\item{outputVar}{- the name of the target column for relabelling}

\item{summarise}{- return dataframe as-is with additional column (FALSE - the default) or return dataframe as group summary with only grouping info and output (TRUE)}
}
\value{
a dbplyr dataframe containing the grouped function
}
\description{
Sometime simpler to rename multiple grouping variables as one single variable which encodes all the possible combinations
}
