% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValues.R
\name{summariseObservations}
\alias{summariseObservations}
\title{summary info about the observations}
\usage{
summariseObservations(df, discreteVars, countVar = NULL)
}
\arguments{
\item{df}{- may be grouped, in which case the value is interpreted as different features}

\item{discreteVars}{- the column(s) of the categorical value (X) quoted by vars(...)}

\item{expected}{- the total number of samples for each outcome expected if there were no missing values.}
}
\description{
summary info about the observations
}
